fix phan quyen

+++ app/Http/Controllers/DanhGiaController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\ChiTietPhanQuyen;
+use App\Models\ChucVu;
 use App\Models\DanhGia;
 use App\Models\HoaDon;
 use App\Models\QuanLyPhim;
@@ -12,23 +14,48 @@ class DanhGiaController extends Controller
 {
     public function getData()
     {
-        $data = DanhGia::leftjoin('khach_hangs', 'khach_hangs.id', 'danh_gias.id_khach_hang')
-            ->leftjoin('quan_ly_phims', 'quan_ly_phims.id', 'danh_gias.id_phim')
-            ->select('danh_gias.*', 'khach_hangs.ten_khach_hang', 'quan_ly_phims.ten_phim')
-            ->get();
-        return response()->json([
-            'data'  =>  $data
-        ]);
+        $id_chuc_nang = 83;
+        $user = Auth::guard('sanctum')->user();
+        $master = ChucVu::where('id', $user->id_chuc_vu)
+            ->first();
+        if ($master->is_master) {
+            $data = DanhGia::leftjoin('khach_hangs', 'khach_hangs.id', 'danh_gias.id_khach_hang')
+                ->leftjoin('quan_ly_phims', 'quan_ly_phims.id', 'danh_gias.id_phim')
+                ->select('danh_gias.*', 'khach_hangs.ten_khach_hang', 'quan_ly_phims.ten_phim')
+                ->get();
+            return response()->json([
+                'data'  =>  $data
+            ]);
+        } else {
+            $check = ChiTietPhanQuyen::join('chuc_vus', 'chuc_vus.id', 'chi_tiet_phan_quyens.id_quyen')
+                ->where('chuc_vus.tinh_trang', 1)
+                ->where('id_quyen', $user->id_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if ($check) {
+                $data = DanhGia::leftjoin('khach_hangs', 'khach_hangs.id', 'danh_gias.id_khach_hang')
+                    ->leftjoin('quan_ly_phims', 'quan_ly_phims.id', 'danh_gias.id_phim')
+                    ->select('danh_gias.*', 'khach_hangs.ten_khach_hang', 'quan_ly_phims.ten_phim')
+                    ->get();
+                return response()->json([
+                    'data'  =>  $data
+                ]);
+            } else {
+                return response()->json([
+                    "message" => 'bạn không có quyền này'
+                ]);
+            }
+        }
     }
     public function getDataChiTietPhim($id)
     {
         try {
             // Using whereRaw with parameters to avoid SQL injection and handle special characters
-        $data = DanhGia::leftjoin('khach_hangs', 'khach_hangs.id', 'danh_gias.id_khach_hang')
-            ->leftjoin('quan_ly_phims', 'quan_ly_phims.id', 'danh_gias.id_phim')
-            ->where('danh_gias.id_phim', $id)
-            ->select('danh_gias.*', 'khach_hangs.ten_khach_hang', 'quan_ly_phims.ten_phim')
-            ->get();
+            $data = DanhGia::leftjoin('khach_hangs', 'khach_hangs.id', 'danh_gias.id_khach_hang')
+                ->leftjoin('quan_ly_phims', 'quan_ly_phims.id', 'danh_gias.id_phim')
+                ->where('danh_gias.id_phim', $id)
+                ->select('danh_gias.*', 'khach_hangs.ten_khach_hang', 'quan_ly_phims.ten_phim')
+                ->get();
 
             return response()->json([
                 'data' => $data
@@ -58,11 +85,11 @@ class DanhGiaController extends Controller
 
         // Kiểm tra khách hàng đã xem phim chưa
         $hoaDon = HoaDon::join('chi_tiet_ves', 'chi_tiet_ves.id_hoa_don', 'hoa_dons.id')
-        ->join('suat_chieus', 'suat_chieus.id', 'chi_tiet_ves.id_suat')
-        ->where('hoa_dons.id_khach_hang', $id_khach_hang)
-        ->where('suat_chieus.phim_id', $id_phim)
-        ->where('hoa_dons.trang_thai', 1)
-        ->first();
+            ->join('suat_chieus', 'suat_chieus.id', 'chi_tiet_ves.id_suat')
+            ->where('hoa_dons.id_khach_hang', $id_khach_hang)
+            ->where('suat_chieus.phim_id', $id_phim)
+            ->where('hoa_dons.trang_thai', 1)
+            ->first();
 
         if (!$hoaDon) {
             return response()->json([
@@ -110,19 +137,43 @@ class DanhGiaController extends Controller
     }
     public function deleteData($id)
     {
-        DanhGia::find($id)->delete();
+        $id_chuc_nang = 84;
+        $user = Auth::guard('sanctum')->user();
+        $master = ChucVu::where('id', $user->id_chuc_vu)
+            ->first();
+        if ($master->is_master) {
+            DanhGia::find($id)->delete();
 
-        return response()->json([
-            'status'    =>  true,
-            'message'   =>  'Đã xoá đánh giá thành công!'
-        ]);
+            return response()->json([
+                'status'    =>  true,
+                'message'   =>  'Đã xoá đánh giá thành công!'
+            ]);
+        } else {
+            $check = ChiTietPhanQuyen::join('chuc_vus', 'chuc_vus.id', 'chi_tiet_phan_quyens.id_quyen')
+                ->where('chuc_vus.tinh_trang', 1)
+                ->where('id_quyen', $user->id_chuc_vu)
+                ->where('id_chuc_nang', $id_chuc_nang)
+                ->first();
+            if ($check) {
+                DanhGia::find($id)->delete();
+
+                return response()->json([
+                    'status'    =>  true,
+                    'message'   =>  'Đã xoá đánh giá thành công!'
+                ]);
+            } else {
+                return response()->json([
+                    "message" => 'bạn không có quyền này'
+                ]);
+            }
+        }
     }
     public function deleteDataClient($id)
     {
         $user = Auth::guard('sanctum')->user();
         DanhGia::where('id_khach_hang', $user->id)
-        ->where('id', $id)
-        ->delete();
+            ->where('id', $id)
+            ->delete();
 
         return response()->json([
             'status'    =>  true,
@@ -134,8 +185,8 @@ class DanhGiaController extends Controller
         $user = Auth::guard('sanctum')->user();
         $data   = $request->all();
         DanhGia::where('id_khach_hang', $user->id)
-        ->where('id', $request->id)
-        ->update($data);
+            ->where('id', $request->id)
+            ->update($data);
         return response()->json([
             'status'    =>  true,
             'message'   =>  'Đã cập nhật đánh giá  thành công!'
